Move bounding rect impl to geo-types
This avoids duplicating its functionality, as it's required in order to make all geometry types and the Geometry enum RStar compatible
