+++ geo-types/src/private_utils.rs
@@ -171,3 +171,38 @@ where
     }
     false
 }
+
+// The Rust standard library has `max` for `Ord`, but not for `PartialOrd`
+pub fn partial_max<T: PartialOrd>(a: T, b: T) -> T {
+    if a > b {
+        a
+    } else {
+        b
+    }
+}
+
+// The Rust standard library has `min` for `Ord`, but not for `PartialOrd`
+pub fn partial_min<T: PartialOrd>(a: T, b: T) -> T {
+    if a < b {
+        a
+    } else {
+        b
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::{partial_max, partial_min};
+
+    #[test]
+    fn test_partial_max() {
+        assert_eq!(5, partial_max(5, 4));
+        assert_eq!(5, partial_max(5, 5));
+    }
+
+    #[test]
+    fn test_partial_min() {
+        assert_eq!(4, partial_min(5, 4));
+        assert_eq!(4, partial_min(4, 4));
+    }
+}
