+++ geo/src/lib.rs
@@ -209,7 +209,10 @@ extern crate rstar;
 pub use crate::algorithm::*;
 pub use crate::types::Closest;
 
-pub use geo_types::{coord, line_string, point, polygon, CoordFloat, CoordNum};
+pub use geo_types::{
+    coord, geometry_cow_delegate_impl, geometry_delegate_impl, line_string, point, polygon,
+    CoordFloat, CoordNum,
+};
 
 pub mod geometry;
 pub use geometry::*;
