+++ geo/src/utils.rs
@@ -74,24 +74,6 @@ where
     }
 }
 
-// The Rust standard library has `max` for `Ord`, but not for `PartialOrd`
-pub fn partial_max<T: PartialOrd>(a: T, b: T) -> T {
-    if a > b {
-        a
-    } else {
-        b
-    }
-}
-
-// The Rust standard library has `min` for `Ord`, but not for `PartialOrd`
-pub fn partial_min<T: PartialOrd>(a: T, b: T) -> T {
-    if a < b {
-        a
-    } else {
-        b
-    }
-}
-
 // Moved to their own module, but we re-export to avoid breaking the API.
 pub use crate::coordinate_position::{coord_pos_relative_to_ring, CoordPos};
 
@@ -153,21 +135,4 @@ pub fn least_and_greatest_index<T: CoordNum>(pts: &[Coord<T>]) -> (usize, usize)
             )
         });
     (min.unwrap().0, max.unwrap().0)
-}
-
-#[cfg(test)]
-mod test {
-    use super::{partial_max, partial_min};
-
-    #[test]
-    fn test_partial_max() {
-        assert_eq!(5, partial_max(5, 4));
-        assert_eq!(5, partial_max(5, 5));
-    }
-
-    #[test]
-    fn test_partial_min() {
-        assert_eq!(4, partial_min(5, 4));
-        assert_eq!(4, partial_min(4, 4));
-    }
 }
